solve UiSecurity integration tests for both Mvc and RazorPages
