+++ demo/GodelTech.Microservices.Security.Demo.Mvc/Controllers/LogoutController.cs
@@ -1,0 +1,16 @@
+﻿using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Mvc;
+
+namespace GodelTech.Microservices.Security.Demo.Mvc.Controllers
+{
+    public class LogoutController : Controller
+    {
+        public async Task<IActionResult> Index()
+        {
+            await HttpContext.SignOutAsync();
+
+            return RedirectToAction("Index", "Home");
+        }
+    }
+}
