+++ demo/GodelTech.Microservices.Security.Demo.Mvc/Controllers/UserController.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace GodelTech.Microservices.Security.Demo.Mvc.Controllers
+{
+    [Authorize]
+    public class UserController : Controller
+    {
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
