+++ demo/GodelTech.Microservices.Security.Demo.Mvc/Views/User/Index.cshtml
@@ -1,0 +1,7 @@
+﻿@{
+    ViewData["Title"] = "User";
+}
+
+<div class="text-center">
+    <h1 class="display-4">@Context.User.Claims.FirstOrDefault(x => x.Type.Equals("name", StringComparison.OrdinalIgnoreCase))?.Value</h1>
+</div>
